add update-only and overwrite options
add additional functionality to updating the database. Either only retrieve seqs for proteins that do not have a seq or whose seq in NCBI has been updated since last retrieval = update-only. overwrite = retrieve protein sequences for all proteins matching user criteria and overwrites the data in the local CAZyme database, regardless if there was a seq there or not, and if the seq had been updated in NCBI since its last retrieval or not